Several enhancements - Add action bar buttons for forms - MainActivity refactoring - Add forms logic refactoring - Minor code formatting fixes
