--- app/src/main/java/br/com/carteira/fragment/AddFiiFormFragment.java
@@ -1,30 +1,0 @@
-package br.com.carteira.fragment;
-
-
-import android.content.Context;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import br.com.carteira.R;
-
-/**
- * A simple {@link Fragment} subclass.
- */
-public class AddFiiFormFragment extends Fragment {
-
-    private Context mContext;
-    private View view;
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        TextView textView = new TextView(getActivity());
-        textView.setText(R.string.hello_blank_fragment);
-        return textView;
-    }
-
-}
