+++ app/src/main/java/br/com/carteira/fragment/BaseAddFormFragment.java
@@ -1,52 +1,36 @@
 package br.com.carteira.fragment;
 
-
-import android.content.Intent;
-import android.support.v4.app.Fragment;
+import android.os.Bundle;
+import android.text.Editable;
+import android.text.TextUtils;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.widget.EditText;
 
-/**
- * A simple {@link Fragment} subclass.
- */
-public class BaseAddFormFragment extends Fragment {
-    // Function that validate the inputted value on the stock symbol field
-    protected Boolean validateStockSymbol(EditText symbol) {
-        // To Do by REGEX
-        if (symbol.getText().toString().length() > 0)
-            return true;
-        else
-            return false;
+import br.com.carteira.R;
+
+public abstract class BaseAddFormFragment extends BaseFragment {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        // Enables the menu
+        setHasOptionsMenu(true);
     }
 
-    // Function that validate the inputted value on the fii symbol field
-    protected Boolean validateFiiSymbol(EditText symbol) {
-        // To Do by REGEX
-        if (symbol.getText().toString().length() > 0)
-            return true;
-        else
-            return false;
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        // Common menu for add forms
+        inflater.inflate(R.menu.add_form_menu, menu);
     }
 
-    // Function that validate the field is not empty on confirm button pressed
-    protected Boolean validateNotEmpty(EditText text) {
-        if (text.getText().toString().length() > 0)
+    // Function that validate if an EditText field is empty
+    protected boolean isEditTextEmpty(EditText symbol) {
+        Editable editable = symbol.getText();
+        if (editable != null && TextUtils.isEmpty(editable.toString())) {
             return true;
-        else
+        } else {
             return false;
-    }
-
-    // Function that validate the date field if the value was inputted correctly
-    protected Boolean validateBuyDate(EditText buyDate) {
-        // To Do by REGEX
-        if (buyDate.getText().toString().length() > 0)
-            return true;
-        else
-            return false;
-    }
-
-    // Sends the information back to the Fragment that called the Dialog and is expecting a result
-    protected void sendResult(int REQUEST_CODE, Intent intent) {
-        getTargetFragment().onActivityResult(
-                getTargetRequestCode(), REQUEST_CODE, intent);
+        }
     }
 }
