+++ app/src/main/java/br/com/carteira/fragment/BaseFragment.java
@@ -1,18 +1,25 @@
 package br.com.carteira.fragment;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 
-/**
- * Created by thipr on 11/9/2016.
- */
+public abstract class BaseFragment extends Fragment {
 
-public class BaseFragment extends Fragment {
+    protected Context mContext;
 
     @Override
-    public void onCreate(Bundle savedInstanceState){
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
         // Keep the fragment alive on screen rotation or other changes
+        //TODO: Is this really needed?
         setRetainInstance(true);
+    }
+
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        mContext = context;
     }
 }
