+++ app/src/main/java/br/com/carteira/fragment/StatisticMainFragment.java
@@ -2,17 +2,14 @@ package br.com.carteira.fragment;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import br.com.carteira.R;
 
-/**
- * A simple {@link Fragment} subclass.
- */
-public class StatisticMainFragment extends Fragment {
+
+public class StatisticMainFragment extends BaseFragment {
 
     public StatisticMainFragment() {
         // Required empty public constructor
