+++ app/src/main/java/br/com/carteira/listener/AddProductListener.java
@@ -1,0 +1,5 @@
+package br.com.carteira.listener;
+
+public interface AddProductListener {
+    void onAddProduct(int productType);
+}
