+++ app/src/main/res/layout/fragment_add_fii_form.xml
@@ -50,20 +50,4 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/buy_date_hint"/>
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal"
-        android:gravity="center">
-        <Button
-            android:id="@+id/cancelButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/cancel_stock_button" />
-        <Button
-            android:id="@+id/buyButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/buy_stock_button" />
-    </LinearLayout>
 </LinearLayout>
