+++ app/src/main/res/menu/add_form_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+      xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/menu_item_done"
+        android:icon="@drawable/ic_menu_done"
+        android:orderInCategory="100"
+        android:title="@string/done"
+        app:showAsAction="ifRoom"/>
+</menu>
